Migrate from uring-sys to newest uring-sys2
uring-sys appears to be not actively developed anymore. `uring-sys2` is
a fork which still receives updates and fixes. In particular, it
includes changes which make it build on musl-based systems.

Unlike `uring-sys`, `uring-sys2` relies fully on bindgen. Thus, a few
seemingly hand-written or altered declrations whch were included in the
former are not availible (at least under the same name) in `uring-sys2`.
In particular, `IoRingOp` is named `io_uring_op` in `uring-sys2`. The
constant `LIBURING_UDATA_TIMEOUT` is gone seemingly without a
replacement. We decided to at least produce the same value.

Some fns used to accept `libc::msghdr`, but `uring-sys2` includes its
own version, which unfortunately is a distinct type. We did compare the
definitions and deemed the layout to be identical.

Other than that, the mutability changed in a few places, but we still
could preserve the existing `iou` interface.